% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_functions.R
\name{add_image}
\alias{add_image}
\title{Add images}
\usage{
add_image(file, position = "middle center", ...)
}
\arguments{
\item{file}{The name of the file in www directory.}

\item{...}{}
}
\value{

}
\description{
Add images
}
