% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_functions.R
\name{add_slider}
\alias{add_slider}
\title{Add text-based sliders}
\usage{
add_slider(inputId, label, choices, position, ...)
}
\arguments{
\item{position...}{}
}
\value{

}
\description{
Wrapper around shinyWidgets sliderTextInput.
}
