% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{add_usernames}
\alias{add_usernames}
\title{Add Usernames}
\usage{
add_usernames(drop_path, num_usernames)
}
\arguments{
\item{drop_path}{The path (folder) in dropbox account where the username file is located}

\item{num_usernames}{The number of usernames to add.}
}
\value{

}
\description{
Add Usernames
}
