% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInformedConsent.R
\name{addInformedConsent}
\alias{addInformedConsent}
\title{Display an informed consent document.}
\usage{
addInformedConsent(file, type)
}
\arguments{
\item{file}{The file (path) containing the informed consent document.}

\item{type}{The file type -- either .docx or PDF.}
}
\value{
A Shiny UI with an informed consent document.
}
\description{
Display an informed consent document.
}
\examples{

\dontrun{
addInformedConsent(file = 'InformedConsent.docx')
}
}
