% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_questions.R
\name{create_survey_questions}
\alias{create_survey_questions}
\title{Create Survey Questions}
\usage{
create_survey_questions()
}
\description{
This function runs a Shiny app that allows the user to create survey
questions for use with the taskdesignr package.
}
