% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_functions.R
\name{define_screen_working}
\alias{define_screen_working}
\title{Define Screen - in Progress}
\usage{
define_screen_working(list_input)
}
\arguments{
\item{list_input}{Takes a list of tags at the moment.}
}
\value{

}
\description{
Define Screen - in Progress
}
