% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDocument.R
\name{demo_displayDocument}
\alias{demo_displayDocument}
\title{Demo for displaying a document in a Shiny app.}
\usage{
demo_displayDocument(file)
}
\arguments{
\item{file}{The file (path) containing a document to display. Can be .docx or .md file type.}
}
\value{
A Shiny App that displays a document.
}
\description{
This function provides a minimal example of using the
\code{\link{displayDocument}} function within your Shiny application to
display a document.
}
\examples{

# FILL THIS OUT LATER
}
