% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_multiple_pages.R
\name{demo_multiple_pages}
\alias{demo_multiple_pages}
\title{Demo Survey over multiple pages}
\usage{
demo_multiple_pages(questions)
}
\arguments{
\item{questions}{A dataframe of questions of which to create a demographic form.}
}
\value{
A Shiny app
}
\description{
This function runs a Shiny app that shows an example of running a demographic survey over multiple pages in Shiny.
Code was inspired by Dean Attali's advanced-shiny Github Repo.
}
\details{
It requires shiny, shinyjs, and shinyWidgets.
}
\examples{

\dontrun{
demo_multiple_pages(teaching_r_questions)
}

}
