% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_single_page.R
\name{demo_single_page}
\alias{demo_single_page}
\title{Demo Survey over multiple pages}
\usage{
demo_single_page(questions)
}
\arguments{
\item{questions}{A dataframe of questions of which to create a demographic form.}
}
\value{
A Shiny App
}
\description{
This function runs a Shiny app that shows an example of running our demographic survey on one page in Shiny.
}
\details{
It requires shiny, shinyjs, and shinyWidgets.
}
\examples{

\dontrun{
demo_single_page(teaching_r_questions)
}
}
