% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDocument.R
\name{displayDocument}
\alias{displayDocument}
\title{Display a document in Shiny.}
\usage{
displayDocument(file)
}
\arguments{
\item{file}{The file (path) containing a document to display.}
}
\value{
A Shiny UI with an informed consent document.
}
\description{
This function will take a Markdown document (.md) or Word document (.docx) and
return the UI for a Shiny app to display it. This should be embedded within a
UI code (see examples for details).
}
\examples{

\dontrun{
ui <- shiny::fluidPage(
shiny::mainPanel(
displayDocument('path-to-your-file')
)
)
}
}
