% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{get_username}
\alias{get_username}
\title{Get a unique username}
\usage{
get_username(group, drop_path)
}
\arguments{
\item{group}{The experimental group. Character strings "A" or "B".}

\item{drop_path}{The path (folder) in dropbox account where the username file is located}
}
\value{

}
\description{
This function takes in a group (either A or B) and a
}
