% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{getServerCode}
\alias{getServerCode}
\title{Server code for adding survey questions}
\usage{
getServerCode(input, df)
}
\arguments{
\item{input}{Input from server}

\item{df}{A user supplied dataframe in the format of teaching_r_questions.}
}
\value{
NA; server code
}
\description{
Create the UI code for a Shiny app based on user-supplied questions. Possible
question (input) types include numeric, text, multiple choice, or selection.
}
