% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{getUICode_individual}
\alias{getUICode_individual}
\title{Generate the UI Code for demoraphic questions}
\usage{
getUICode_individual(df)
}
\arguments{
\item{df}{A nested dataframe.}
}
\value{
UI Code for a Shiny App.
}
\description{
Generate the UI Code for demoraphic questions
}
