% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_run.R
\name{multiply_trials}
\alias{multiply_trials}
\title{Define a task run}
\usage{
multiply_trials(trial_temps, num_trials)
}
\arguments{
\item{trial_temps}{The trial templates}

\item{num_trials}{The number of trials in a run}
}
\value{

}
\description{
Define a task run
}
