% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_functions.R
\name{place_horizontal}
\alias{place_horizontal}
\title{Place UI elements in the horizontal position for trial screens}
\usage{
place_horizontal(object)
}
\arguments{
\item{object}{The output of \code{\link{add_text}} and other taskdesignr functions.}
}
\value{
Positioned HTML tag.
}
\description{
Place UI elements in the horizontal position for trial screens
}
