% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{rand_str}
\alias{rand_str}
\title{Generate Random Strings from random.org}
\usage{
rand_str(
  n = 10,
  len = 6,
  digits = TRUE,
  upperalpha = TRUE,
  loweralpha = TRUE,
  unique = TRUE
)
}
\arguments{
\item{n}{Number of strings}

\item{len}{Length of strings}

\item{digits}{Logical: TRUE and digits will be in string, FALSE and they won't.}

\item{upperalpha}{Logical: TRUE and capital letters will be in string, FALSE and they won't.}

\item{loweralpha}{Logical: TRUE and lowercase will be in string, FALSE and they won't.}

\item{unique}{Logical: TRUE and strings will be unique, FALSE and they won't.}
}
\value{
A data frame with column name "strings" containing n unique strings.
}
\description{
This function is minimally adapted from the CRAN package "random" which
provides access to random.org.
}
\examples{

strings <- rand_str(upperalpha = FALSE)
}
