% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{rand_str_update}
\alias{rand_str_update}
\title{Generate Random Strings}
\usage{
rand_str_update(
  n = 10,
  length = 6,
  digits = TRUE,
  upperalpha = TRUE,
  loweralpha = TRUE,
  unique = TRUE
)
}
\arguments{
\item{n}{Number of strings}

\item{digits}{Logical: TRUE and digits will be in string, FALSE and they won't.}

\item{upperalpha}{Logical: TRUE and capital letters will be in string, FALSE and they won't.}

\item{loweralpha}{Logical: TRUE and lowercase will be in string, FALSE and they won't.}

\item{unique}{Logical: TRUE and strings will be unique, FALSE and they won't.}

\item{len}{Length of strings}
}
\value{

}
\description{
Generate n random strings of any length using a combination of uppercase or
lowercase letters and digits 0-9.
}
\examples{

user_ids <- data.frame(id = rand_str_update(n = 50))

}
