% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_screen.R
\name{screen_template}
\alias{screen_template}
\title{Define templates for individual screens}
\usage{
screen_template(screen_type, screen_duration, ui)
}
\arguments{
\item{screen_type}{A string that describes the component of the trial this
applies to. For example, "presentation", "choice", or "outcome".}

\item{screen_duration}{A time (milliseconds) that this screen will be
displayed for. Users may also supply a function that returns a numeric
(e.g. rpois(1,6000)).}

\item{ui}{A shiny UI object or definition}
}
\value{
A list defining the template for the screen type.
}
\description{
This function defines what occurs on the individual screens within a trial.
}
