% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{setup_usernames}
\alias{setup_usernames}
\title{Set up the username files on Dropbox}
\usage{
setup_usernames(drop_path, num_usernames)
}
\arguments{
\item{drop_path}{The path (folder) in dropbox account where the username file is located}

\item{num_usernames}{Number of usernames to generate between both groups. Must be divisible by two.}
}
\value{

}
\description{
Set up the username files on Dropbox
}
\examples{

}
