% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{surveyUI}
\alias{surveyUI}
\title{Generate the UI Code for demoraphic questions}
\usage{
surveyUI(df)
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}
}
\value{
UI Code for a Shiny App.
}
\description{
Create the UI code for a Shiny app based on user-supplied questions. Possible
question (input) types include numeric, text, multiple choice, or selection.
}
\examples{
\dontrun{
surveyUI(teaching_r_questions)
}
}
