% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teaching_r_questions}
\alias{teaching_r_questions}
\title{A sample CSV file for demographic questions}
\format{
A data frame with 54 rows and 6 columns:
\describe{
\itemize{
\item question
\item option
\item input_type
\item label
\item dependence
\item dependence_value
}
}
}
\source{
\url{D'Agostino McGowan Data Science Lab at Wake Forest University}
}
\usage{
teaching_r_questions
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
