% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_trial.R
\name{trial_template}
\alias{trial_template}
\title{Define a template for trials}
\usage{
trial_template(screen_temps)
}
\arguments{
\item{screen_temps}{A list of screen templates defined with \code{\link{screen_template}}.}
}
\value{

}
\description{
Define a template for trials
}
