% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{update_username_vector}
\alias{update_username_vector}
\title{Update username vector on Dropbox}
\usage{
update_username_vector(user_vec, file, drop_path)
}
\arguments{
\item{user_vec}{The last used vector}

\item{file}{The username file on dropbox.}

\item{drop_path}{The path (folder) in dropbox account where the username file is located}
}
\value{

}
\description{
This function is called on exit when supplying a new username for the
teaching r study. It takes in the vector most recently used to supply a
username, removes the entry that was used, and reuploads the file.
}
