% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_user_ids.R
\name{remove_usernames}
\alias{remove_usernames}
\title{Remove usernames}
\usage{
remove_usernames(drop_path, num_usernames)
}
\arguments{
\item{num_usernames}{The number of usernames to remove.}
}
\value{

}
\description{
Remove usernames
}
